requirements.txt
asgiref==3.3.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
CacheControl==0.12.6
cachetools==4.2.2
certifi==2020.11.8
cffi==1.14.5
channels==3.0.4
chardet==3.0.4
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.1.3
djangorestframework==3.12.2
firebase-admin==5.0.1
google-api-core==1.30.0
google-api-python-client==2.11.0
google-auth==1.32.1
google-auth-httplib2==0.1.0
google-cloud-core==1.7.1
google-cloud-firestore==2.1.3
google-cloud-storage==1.40.0
google-crc32c==1.1.2
google-resumable-media==1.3.1
googleapis-common-protos==1.53.0
grpcio==1.38.1
httplib2==0.19.1
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
msgpack==1.0.2
packaging==20.9
Pillow==8.0.1
proto-plus==1.19.0
protobuf==3.17.3
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pytz==2020.4
requests==2.25.0
rsa==4.7.2
service-identity==21.1.0
six==1.15.0
sqlparse==0.4.1
twilio==6.48.0
Twisted==21.7.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.10.0.0
uritemplate==3.0.1
urllib3==1.26.2
zope.interface==5.4.0
